+++ tests/EncodeTest.php
@@ -11,6 +11,7 @@ use PHPUnit\Framework\TestCase;
 use SandFox\Bencode\Bencode;
 use SandFox\Bencode\Exceptions\InvalidArgumentException;
 use SandFox\Bencode\Types\BencodeSerializable;
+use SandFox\Bencode\Types\DictType;
 use SandFox\Bencode\Types\ListType;
 use stdClass;
 
@@ -137,6 +138,9 @@ class EncodeTest extends TestCase
 
         self::assertEquals('d3:key5:value4:test8:whatevere', Bencode::encode(
             new ArrayObject(['key' => 'value', 'test' => 'whatever'])
+        ));
+        self::assertEquals('d3:key5:value4:test8:whatevere', Bencode::encode(
+            new DictType(new ArrayObject(['key' => 'value', 'test' => 'whatever']))
         ));
 
         // even sequential