+++ tests/EncodeEmptyTest.php
@@ -46,7 +46,7 @@ class EncodeEmptyTest extends TestCase
         $this->expectExceptionMessage('Unable to encode an empty value');
 
         Bencode::encode(new class implements BencodeSerializable {
-            public function bencodeSerialize(): mixed
+            public function bencodeSerialize(): bool
             {
                 return false;
             }