+++ tests/LargeIntegerTest.php
@@ -104,6 +104,11 @@ class LargeIntegerTest extends TestCase
         $decodedNeg = Bencode::decode($encodedNeg, bigInt: Bencode\BigInt::GMP);
         self::assertInstanceOf(\GMP::class, $decoded);
         self::assertEquals($expectedNeg, $decodedNeg);
+
+        // legacy option
+        $decodedNeg = Bencode::decode($encodedNeg, ['useGMP' => true]);
+        self::assertInstanceOf(\GMP::class, $decoded);
+        self::assertEquals($expectedNeg, $decodedNeg);
     }
 
     // Pear