+++ src/Bencode.php
@@ -89,23 +89,34 @@ final class Bencode
      * Encode arbitrary data to bencode string
      *
      * @param mixed $data
+     * @param array $options
+     * @param bool $useJsonSerializable
      * @return string
      */
-    public static function encode(mixed $data): string
-    {
-        return (new Encoder())->encode($data);
+    public static function encode(
+        mixed $data,
+        array $options = [],
+        bool $useJsonSerializable = false,
+    ): string {
+        return (new Encoder($options, $useJsonSerializable))->encode($data);
     }
 
     /**
      * Dump data to bencoded stream
      *
      * @param mixed $data
-     * @param resource|null $writeStream Write capable stream. If null, a new php://temp will be created
+     * @param null $writeStream Write capable stream. If null, a new php://temp will be created
+     * @param array $options
+     * @param bool $useJsonSerializable
      * @return resource Original or created stream
      */
-    public static function encodeToStream(mixed $data, $writeStream = null)
-    {
-        return (new Encoder())->encodeToStream($data, $writeStream);
+    public static function encodeToStream(
+        mixed $data,
+        $writeStream = null,
+        array $options = [],
+        bool $useJsonSerializable = false,
+    ) {
+        return (new Encoder($options, $useJsonSerializable))->encodeToStream($data, $writeStream);
     }
 
     /**
@@ -113,10 +124,16 @@ final class Bencode
      *
      * @param string $filename
      * @param mixed $data
+     * @param array $options
+     * @param bool $useJsonSerializable
      * @return bool success of file_put_contents
      */
-    public static function dump(string $filename, mixed $data): bool
-    {
-        return (new Encoder())->dump($data, $filename);
+    public static function dump(
+        string $filename,
+        mixed $data,
+        array $options = [],
+        bool $useJsonSerializable = false,
+    ): bool {
+        return (new Encoder($options, $useJsonSerializable))->dump($data, $filename);
     }
 }