+++ CHANGELOG.md
@@ -18,7 +18,8 @@
     * Stringables no longer become strings by default.
       Use ``useStringable: true`` to return old behavior.
     * ``dump($filename, $data)`` became ``dump($data, $filename)`` for consistency with streams.
-* `Decoder` and `Encoder` objects that can be pre-configured and then used with consistent options. 
+* `Decoder` and `Encoder` objects that can be pre-configured and then used with consistent options.
+* `bencodeSerialize` now declares `mixed` return type
 
 ## 2.x