+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 2.x
 
+### 2.next
+
+* Decoder and Encoder are backported from 3.x
+* `useGMP` is marked as deprecated
+* Fixed `'useGMP'` in options array causing crash
+
 ### 2.6.1
 
 *Sep 10, 2021*