+++ CHANGELOG.md
@@ -74,6 +74,8 @@
 ### 1.next
 
 * Decoder and Encoder are backported from 3.x
+* `DictType` backported from 3.x
+* `useJsonSerializable` backported from 3.x
 
 ### 1.6.2