+++ CHANGELOG.md
@@ -73,7 +73,9 @@
 
 ## 1.x
 
-### 1.next
+### 1.7.0
+
+*Sep 17, 2021*
 
 * Decoder and Encoder are backported from 3.x
 * `DictType` backported from 3.x